Block editor: move block tools components and styles to separate folder
