+++ packages/block-editor/src/components/block-list/use-in-between-inserter.js
@@ -10,7 +10,7 @@ import { useContext } from '@wordpress/element';
  * Internal dependencies
  */
 import { store as blockEditorStore } from '../../store';
-import { InsertionPointOpenRef } from './insertion-point';
+import { InsertionPointOpenRef } from '../block-tools/insertion-point';
 
 export function useInBetweenInserter() {
 	const openRef = useContext( InsertionPointOpenRef );
