+++ packages/block-editor/src/components/block-tools/block-selection-button.js
