+++ packages/block-editor/src/components/block-tools/index.js
@@ -1,0 +1,14 @@
+/**
+ * Internal dependencies
+ */
+import InsertionPoint from './insertion-point';
+import BlockPopover from './block-popover';
+
+export default function BlockTools( { children } ) {
+	return (
+		<InsertionPoint>
+			<BlockPopover />
+			{ children }
+		</InsertionPoint>
+	);
+}
