+++ packages/block-editor/src/style.scss
@@ -10,6 +10,7 @@
 @import "./components/block-icon/style.scss";
 @import "./components/block-inspector/style.scss";
 @import "./components/block-list/style.scss";
+@import "./components/block-tools/style.scss";
 @import "./components/block-list-appender/style.scss";
 @import "./components/block-breadcrumb/style.scss";
 @import "./components/block-card/style.scss";
